src/main/java/com/fusion/parser/EmailExtractor.java
package com.fusion.parser;

import java.util.regex.Matcher;
import java.util.regex.Pattern;

import org.springframework.stereotype.Component;

@Component
public class EmailExtractor {
	private Pattern pattern;
	private Matcher matcher;

	/*private static final String EMAIL_PATTERN =
		"^[_A-Za-z0-9-\\+]+(\\.[_A-Za-z0-9-]+)*@"
		+ "[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$";
	*/	
		
	
	private static final String EMAIL_PATTERN = "[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+";
	
	public EmailExtractor() {
		pattern = Pattern.compile("[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+");
	}

	/**
	 * Validate hex with regular expression
	 *
	 * @param hex
	 *            hex for validation
	 * @return true valid hex, false invalid hex
	 */
	public String extract(final String hex) {
		String email = null;
		Matcher matcher = pattern.matcher(hex);
	    while (matcher.find()) {
	        email = matcher.group();
	    }
	    return email;
	}
}
